requirements.txt
keyboard>=0.13.5
Pillow>=9.0.0
requests>=2.28.0
